cloud: make `make_tls_config_for_scylla_cloud_host` method on CloudConfig
It makes great sense, because the CloudConfig is the subject of the
operation.
