+++ scylla/src/cloud/config.rs
@@ -1,9 +1,13 @@
+use std::net::SocketAddr;
 use std::{collections::HashMap, io};
 
 use scylla_cql::{frame::types::SerialConsistency, Consistency};
 use thiserror::Error;
+use tracing::warn;
+use uuid::Uuid;
 
-use crate::network::TlsError;
+use crate::client::session::TlsContext;
+use crate::network::{TlsConfig, TlsError};
 
 #[non_exhaustive]
 #[derive(Debug, Error)]
@@ -65,6 +69,48 @@ impl CloudConfig {
         self.auth_infos.get(auth_info_name).expect(
             "BUG: Validation should have prevented current context's auth info pointing to unknown auth_info",
         )
+    }
+
+    pub(crate) fn make_tls_config_for_scylla_cloud_host(
+        &self,
+        host_id: Option<Uuid>,
+        dc: Option<&str>,
+        proxy_address: SocketAddr,
+    ) -> Result<Option<TlsConfig>, TlsError> {
+        let Some(datacenter) = dc.and_then(|dc| self.get_datacenters().get(dc)) else {
+            warn!("Datacenter {:?} of node {:?} with addr {} not described in cloud config. Proceeding without setting SNI for the node, which will most probably result in nonworking connections,.",
+                   dc, host_id, proxy_address);
+            // FIXME: Consider returning error here.
+            return Ok(None);
+        };
+
+        let domain_name = datacenter.get_node_domain();
+        let auth_info = self.get_current_auth_info();
+
+        let tls_context = match auth_info.get_tls() {
+            #[cfg(feature = "openssl-010")]
+            TlsInfo::OpenSsl010 { key, cert } => {
+                use openssl::ssl::{SslContext, SslMethod, SslVerifyMode};
+                let mut builder = SslContext::builder(SslMethod::tls())?;
+                builder.set_verify(if datacenter.get_insecure_skip_tls_verify() {
+                    SslVerifyMode::NONE
+                } else {
+                    SslVerifyMode::PEER
+                });
+                let ca = datacenter.openssl_ca();
+                builder.cert_store_mut().add_cert(ca.clone())?;
+                builder.set_certificate(cert)?;
+                builder.set_private_key(key)?;
+                let context = builder.build();
+                TlsContext::OpenSsl010(context)
+            }
+        };
+
+        Ok(Some(TlsConfig::new_for_sni(
+            tls_context,
+            domain_name,
+            host_id,
+        )))
     }
 }
